// api/src/types.ts
export interface ApiItem {
    anywaySelling: boolean;
    brand: string;
    offerDescription: string;
    fullPrice: number;
    grayDescription: string;
    highlight: boolean;
    id: string;
    isGeneric: boolean;
    mediaDescription: string;
    largeDescription: string;
    mediaImageUrl: string;
    offerPrice: number;
    outstanding: boolean;
    requirePrescription: string;
    sales: string;
    spaces: string;
    status: string;
    taxRate: number;
    totalStock: number;
    quantitySold: number;
    idClassification: number;
    unitPrice: number;
    taxes: number;
    listUrlImages?: string[];
    categorie: string;
    marca: string;
    onlyOnline: boolean;
    departments: string[];
    subCategory: string;
    supplier: string;
    availableOnline: boolean;
    noFullPrice: boolean;
    allowedStock: number;
    globalStock: number;
    deliveryPrice: number;
    greenDelivery: boolean;
    greenStore: boolean;
    offerStartDate: string;
    offerEndDate: string;
    outofstore: boolean;
    measurePum: number;
    labelPum: string;
    requirePrescriptionMedical: boolean;
    flashOffer: boolean;
    offerText?: string;
}